+++ src/components/main/film-card/film-card.tsx
@@ -1,6 +1,6 @@
-import { useState } from 'react';
+import { memo, useState } from 'react';
 import { Link } from 'react-router-dom';
-import DefaultPlayer from '../../default-player/default-player';
+import MiniPlayer from '../../mini-player/mini-player';
 import { filmCardVideoProps } from '../../../const';
 import { Film } from '../../../types/film';
 
@@ -8,7 +8,7 @@ type FilmCardProps = {
   film: Film;
 }
 
-export default function FilmCard({ film }: FilmCardProps): JSX.Element {
+function FilmCard({ film }: FilmCardProps): JSX.Element {
   const [isActive, changeActive] = useState(false);
 
   return (
@@ -20,12 +20,11 @@ export default function FilmCard({ film }: FilmCardProps): JSX.Element {
       <Link
         className="small-film-card__link"
         to={`/films/${film.id}`}
-        // reloadDocument
       >
         <div className="small-film-card__image">
           {
             isActive ?
-              <DefaultPlayer {...{ ...filmCardVideoProps, ...film, isActive }} /> :
+              <MiniPlayer {...{ ...filmCardVideoProps, ...film, isActive }} /> :
               <img src={film.previewImage} alt={film.name} width="280" height="175" />
           }
         </div>
@@ -36,3 +35,5 @@ export default function FilmCard({ film }: FilmCardProps): JSX.Element {
     </article>
   );
 }
+
+export default memo(FilmCard);
