+++ src/types/genre.ts
@@ -1,0 +1,4 @@
+export type Genre = {
+    id: number;
+    title: string;
+};
